✨ StampDetectResult 추가 완료
